Respect total and count max items correctly
The number of items read was not limited to "total" (like a "tail"
would do), but to the lowest multiple of "max items" that is greater
than or equal to total.

Also, the displayed count assumed that you always got "max items",
as long as there was at least that many items.

So we count the number of items read, use that when tallying the
number of items read, and decrease total by that number so that
we can limit the max items when it's greater than the (remaining)
total.

There's no test case for correct counting when less than asked for
records are returned.