Adopt error_chain for better error reporting
Note that jq_rs::Error cannot be used seamlessly because it doesn't
implement Send, and I can't implement it for it, as far as I know.

The work-around was rather more difficult to implement than I ever
expected, because error_chain doesn't make it easy to _create_
errors. The work-around relies on a separate method to create a
Result with the error, and then unpack it so it can be used on a
map_err, so that jq's error can be _replaced_ instead of chained.