Use jq-rs bundled feature on linux builds
The bundled feature adds jq-src as a dependency, which downloads
and compiles jq 1.6. Ubuntu does not have jq 1.6 available yet,
and jq-rs has bindings for jq 1.6.

The per-target configuration of dependencies does not actually
work at the moment, so, unfortunately, MacOS users will have to
put up with the extra build time, and install autotools, etc.