Add parallel execution [ch8673]
Split main logic into functions so that it can be run in background,
as add a configurable number of workers.

This change introduces use of temporary files to store the decoded result
of each worker. This allows the output of the workers to be unified. An
alternate implementation that skewed use of temporary files but relied
on reading the result of each worker line by line was tested and found
to have subpar performance negating all parallelism gains. The key
elements of the logic can be seen in the way the name of the temporary
files are passed to the main thread.

An undocumented option was added to allow read input items from a file
instead of going to dynamodb, for ease of testing.

The processing function was split off as well, for reasons that
disappeared from the version that was finally implemented, but it also
avoids duplication implementing the previous option, and it should help
implement an option to automatically replace that loop with the native
client in the future. YAGNI would apply, but since it _was_ refactored,
it would be _more_ work to inline it again.