Add sanity check/recovery on temporary file names
Despite the race condition code for incomplete file names, we are
still getting errors where the temporary file names are missing the
first slash. Also, I've never seen this error manifest in any other
way than a missing first slash.

So, if it's just the first slash, this code will work around it. If
we are getting a bad file name to begin with, we'll complain about it
and exit, to give us a better diagnostic.